streamlit/requirements.txt
pillow
requests
torch 
torchvision
streamlit

json5
numpy
pandas
python-dotenv
streamlit